Cache depthDataFormat in Depth and move GPU depth conversion logic to GPUDepthConverter
- Add session-wide depthDataFormat caching to Depth.ts.
- Refactor WebGL GPU-to-CPU depth conversion into src/depth/GPUDepthConverter.ts.
- Simplify DepthMesh to strictly accept CPU depth, removing all conversion properties and logic.
- Add updateFullResolutionDepthMesh() parameterless helper to Depth.ts.
- Update demos/splash/SplashScript.js to call updateFullResolutionDepthMesh().
